# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "version": "1.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Lista de tareas"
                ],
                "summary": "Mostrar lista de tareas",
                "operationId": "a1e508a15d231d9b7c68e16032a7952e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tareas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Leer"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Leer documentación de proyecto."
                                                    },
                                                    "completed": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "due_date": {
                                                        "type": "string",
                                                        "example": "2024-11-23"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-11-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-11-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Registrar tarea"
                ],
                "summary": "Registrar la información de una Tarea",
                "operationId": "579b0cbd3e383b04fdf5c03a61e24291",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "completed": {
                                                "type": "boolean"
                                            },
                                            "due_date": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Leer",
                                    "description": "Leer documentación de proyecto.",
                                    "completed": 1,
                                    "due_date": "2024-11-23"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tarea": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Leer"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Leer documentación de proyecto."
                                                },
                                                "completed": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "due_date": {
                                                    "type": "string",
                                                    "example": "2024-11-23"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-11-23T00:09:16.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-11-23T12:33:45.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de los datos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title field is required."
                                                    }
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The description field is required."
                                                    }
                                                },
                                                "completed": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The completed field is required."
                                                    }
                                                },
                                                "due_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The due date field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}